+++ CHANGELOG.md
@@ -5,6 +5,21 @@ Changelog
 Unreleased
 ----------
 
+3.1.0 (2026-04-29)
+------------------
+
+### Deprecated
+
+ - Legacy content elements `bs_accordion_start`, `bs_accordion_end`,
+   `bs_accordion_group_start` and `bs_accordion_group_end` are deprecated.
+   Use `bs_accordion_wrapper` and `bs_accordion_group_wrapper` instead.
+
+### Added
+
+ - Bundle configuration option `enable_legacy_elements` (default: `true`) to
+   disable legacy content elements.
+ - Add new content element `bs_accordion_wrapper` and `bs_accordion_group_wrapper`. Enable migration with bundle configuration.
+
 3.0.0
 -----